separated out test cases to relevant components
